backend/pyproject.toml
[project]
name = "solalex"
version = "0.1.0-beta.0"
description = "Solalex backend — sekundengenaue Nulleinspeisung und Akku-Pool-Steuerung."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "Proprietary" }
authors = [{ name = "ALKLY", email = "info@alkly.de" }]

dependencies = [
    "fastapi[standard]>=0.135.1",
    "uvicorn[standard]>=0.30",
    "aiosqlite>=0.20",
    "websockets>=13",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/solalex"]

# ----- Ruff -----
[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # simplify
    "ASYNC",
]
ignore = ["E501"]  # line-length handled by formatter

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["N802", "N803"]  # test funcs can use non-PEP8 names

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ----- Mypy -----
[tool.mypy]
python_version = "3.13"
strict = true
files = ["src/solalex", "tests"]
disallow_untyped_decorators = false
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["aiosqlite.*"]
ignore_missing_imports = true

# ----- Pytest -----
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
